requirements.txt
Flask==1.1.2
numpy==1.21.5
pandas==1.4.2
python_dateutil==2.8.2
requests==2.27.1
yfinance==0.1.87
